Simplified task orchestration and aligned task namespaces with folder layout
- Replaced the Pipeline class with a private RunTask helper on AnalysisPipeline (every call site was new Pipeline(singleTask).Execute(ctx) — the multi-task plumbing earned no keep); deleted OspreySharp.Tasks/Pipeline.cs
- Consolidated four near-identical PipelineContext constructions in AnalysisPipeline.Run into a single shared context, matching the design intent already documented on PipelineContext
- Moved all four concrete task classes into namespace pwiz.OspreySharp.Tasks (was per-task sub-namespaces); cleaned up the redundant usings + doc-comment crefs that ReSharper flagged across the task files + FileSaver
- OspreySharp.Tasks now exposes just OspreyTask + PipelineContext — the shared framework boundary the four task subclasses use
- Snapshot regression PASS at every stage on Stellar 3-file; 303/303 OspreySharp unit tests pass

See ai/todos/active/TODO-20260509_osprey_pipeline_tasks.md
