+++ pwiz_tools/OspreySharp/OspreySharp.IO/FileSaver.cs
@@ -18,7 +18,6 @@
  */
 
 using System;
-using System.Diagnostics;
 using System.IO;
 using System.Runtime.InteropServices;
 using System.Text;
@@ -64,7 +63,7 @@ namespace pwiz.OspreySharp.IO
         /// Resolves <paramref name="fileName"/> against the current
         /// working directory so a bare-filename argument (no directory
         /// component) lands the temp file alongside its destination
-        /// rather than failing inside <see cref="Path.GetDirectoryName"/>
+        /// rather than failing inside <see cref="Path.GetDirectoryName(System.ReadOnlySpan{char})"/>
         /// downstream. Diverges from the SharedBatch original which
         /// happens to never see relative paths.
         /// </summary>
